+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/room/settings/RoomSettingsNotifications.kt
@@ -1,11 +1,8 @@
 package de.connect2x.messenger.compose.view.room.settings
 
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.NotificationImportant
@@ -13,9 +10,7 @@ import androidx.compose.material.icons.filled.Notifications
 import androidx.compose.material.icons.filled.NotificationsActive
 import androidx.compose.material.icons.filled.NotificationsOff
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.RadioButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
@@ -25,9 +20,12 @@ import androidx.compose.ui.text.capitalize
 import androidx.compose.ui.text.intl.Locale
 import androidx.compose.ui.unit.dp
 import de.connect2x.messenger.compose.view.DI
+import de.connect2x.messenger.compose.view.Tooltip
+import de.connect2x.messenger.compose.view.common.RadioSetting
+import de.connect2x.messenger.compose.view.common.RadioSettingOption
+import de.connect2x.messenger.compose.view.common.TooltipText
 import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.i18n.I18nView
-import de.connect2x.trixnity.messenger.viewmodel.room.settings.NotificationLevel
 import de.connect2x.trixnity.messenger.viewmodel.room.settings.NotificationLevels
 import de.connect2x.trixnity.messenger.viewmodel.room.settings.RoomSettingsNotificationsViewModel
 
@@ -46,6 +44,14 @@ class RoomSettingsNotificationsViewImpl : RoomSettingsNotificationsView {
     override fun create(roomSettingsNotificationsViewModel: RoomSettingsNotificationsViewModel) {
         val i18n = DI.get<I18nView>()
         val isLoading = roomSettingsNotificationsViewModel.isNotificationsLevelLoading.collectAsState().value
+        val selectedLevel = roomSettingsNotificationsViewModel.selectedRoomNotificationsLevel.collectAsState().value
+        val selectedLevelName = selectedLevel.name.collectAsState().value
+        val selectedLevelExplanation = selectedLevel.explanation.collectAsState().value
+        val roomNotificationLevels = roomSettingsNotificationsViewModel.roomNotificationLevels
+        val roomNotificationLevelDefault = roomNotificationLevels.getValue(NotificationLevels.DEFAULT)
+        val roomNotificationLevelAll = roomNotificationLevels.getValue(NotificationLevels.ALL)
+        val roomNotificationLevelMentions = roomNotificationLevels.getValue(NotificationLevels.MENTIONS)
+        val roomNotificationLevelSilent = roomNotificationLevels.getValue(NotificationLevels.SILENT)
 
         Column {
             Row(verticalAlignment = Alignment.CenterVertically) {
@@ -53,70 +59,45 @@ class RoomSettingsNotificationsViewImpl : RoomSettingsNotificationsView {
                     text = i18n.commonNotifications().capitalize(Locale.current),
                     style = MaterialTheme.typography.titleMedium,
                 )
-                if (isLoading) {
-                    Spacer(Modifier.size(20.dp))
-                    CircularProgressIndicator(Modifier.size(20.dp), strokeWidth = 2.dp)
-                }
             }
-            RoomSettingsNotificationsLevel(
-                roomSettingsNotificationsViewModel,
-                roomSettingsNotificationsViewModel.roomNotificationLevels.getValue(NotificationLevels.DEFAULT)
-            )
-            RoomSettingsNotificationsLevel(
-                roomSettingsNotificationsViewModel,
-                roomSettingsNotificationsViewModel.roomNotificationLevels.getValue(NotificationLevels.ALL)
-            )
-            RoomSettingsNotificationsLevel(
-                roomSettingsNotificationsViewModel,
-                roomSettingsNotificationsViewModel.roomNotificationLevels.getValue(NotificationLevels.MENTIONS)
-            )
-            RoomSettingsNotificationsLevel(
-                roomSettingsNotificationsViewModel,
-                roomSettingsNotificationsViewModel.roomNotificationLevels.getValue(NotificationLevels.SILENT)
+            Spacer(Modifier.size(10.dp))
+            RadioSetting(
+                title = {
+                    if (isLoading) {
+                        CircularProgressIndicator(Modifier.size(20.dp), strokeWidth = 2.dp)
+                    } else {
+                        Tooltip(tooltip = { TooltipText { selectedLevelExplanation } }) {
+                            Text(selectedLevelName)
+                        }
+                    }
+                },
+                options = mapOf(
+                    roomNotificationLevelDefault to RadioSettingOption(
+                        roomNotificationLevelDefault.name.collectAsState().value,
+                        roomNotificationLevelDefault.explanation.collectAsState().value
+                    ),
+                    roomNotificationLevelAll to RadioSettingOption(
+                        roomNotificationLevelAll.name.collectAsState().value,
+                        roomNotificationLevelAll.explanation.collectAsState().value
+                    ),
+                    roomNotificationLevelMentions to RadioSettingOption(
+                        roomNotificationLevelMentions.name.collectAsState().value,
+                        roomNotificationLevelMentions.explanation.collectAsState().value
+                    ),
+                    roomNotificationLevelSilent to RadioSettingOption(
+                        roomNotificationLevelSilent.name.collectAsState().value,
+                        roomNotificationLevelSilent.explanation.collectAsState().value
+                    )
+                ),
+                value = selectedLevel,
+                set = { roomSettingsNotificationsViewModel.changeSelectedRoomNotificationsLevel(newLevel = it) },
+                icon = when (selectedLevel.key) {
+                    NotificationLevels.DEFAULT -> Icons.Default.Notifications
+                    NotificationLevels.ALL -> Icons.Default.NotificationsActive
+                    NotificationLevels.MENTIONS -> Icons.Default.NotificationImportant
+                    NotificationLevels.SILENT -> Icons.Default.NotificationsOff
+                }
             )
         }
-    }
-}
-
-@Composable
-fun RoomSettingsNotificationsLevel(
-    roomSettingsNotificationsViewModel: RoomSettingsNotificationsViewModel,
-    level: NotificationLevel,
-) {
-    val i18n = DI.get<I18nView>()
-    val selectedRoomNotificationLevel =
-        roomSettingsNotificationsViewModel.selectedRoomNotificationsLevel.collectAsState()
-    val name = level.name.collectAsState()
-    val explanation = level.explanation.collectAsState()
-
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        modifier = Modifier.clickable {
-            roomSettingsNotificationsViewModel.changeSelectedRoomNotificationsLevel(
-                level
-            )
-        }
-    ) {
-        RadioButton(
-            selected = selectedRoomNotificationLevel.value.key == level.key,
-            onClick = { roomSettingsNotificationsViewModel.changeSelectedRoomNotificationsLevel(level) },
-        )
-        when (level.key) {
-            NotificationLevels.DEFAULT -> Icon(Icons.Default.Notifications, i18n.commonDefault())
-            NotificationLevels.ALL -> Icon(Icons.Default.NotificationsActive, i18n.commonOn())
-            NotificationLevels.MENTIONS -> Icon(Icons.Default.NotificationImportant, i18n.roomSettingsMentions())
-            NotificationLevels.SILENT -> Icon(Icons.Default.NotificationsOff, i18n.commonOff())
-        }
-        Spacer(Modifier.size(10.dp))
-        Text(
-            text = name.value,
-            style = MaterialTheme.typography.labelLarge
-        )
-    }
-    Box(Modifier.padding(start = 48.dp)) {
-        Text(
-            text = explanation.value,
-            style = MaterialTheme.typography.bodyMedium,
-        )
     }
 }