+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Insert formula in padding of the timeline
 - Increase size of avatar and change order of settings in profile view
 - Check for server reachability when logout fails
+- Make room notification settings collapsable and change setting design to the one used by other room settings
+- Disable notification setting sub-options when notifications are disabled
 
 ### Deprecated